MultitenancyProperties

+++ src/main/java/lt/jauga/service/PostService.java
@@ -8,18 +8,12 @@ import java.util.Optional;
 
 public interface PostService {
 
-    Optional<Post> findForId(Long id);
+    Optional<Post> findForId(int id);
 
     Post save(Post post);
 
-    /**
-     * Finds a {@link Page) of {@link Post} of provided user ordered by date
-     */
     Page<Post> findByUserOrderedByDatePageable(User user, int page);
 
-    /**
-     * Finds a {@link Page) of all {@link Post} ordered by date
-     */
     Page<Post> findAllOrderedByDatePageable(int page);
 
     void delete(Post post);
